fix(help、consultant):修改参数及consultant部分逻辑

+++ src/main/java/com/example/heart_field/dto/consultant/profile/ConsultantProfileDto.java
@@ -12,6 +12,6 @@ public class ConsultantProfileDto {
     private Integer maxCount;
     private Integer maxConcurrentCount;
     private Integer currentSessionCount;
-    private Long todayTotalTime;
+    private Integer todayTotalTime;
     private Integer todayTotalCount;
 }
